4 make some routes

+++ app/forms/sound_form.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import (StringField, DecimalField, IntegerField, BooleanField, BooleanField)
+from wtforms import (StringField, IntegerField, BooleanField, BooleanField)
 from wtforms.validators import DataRequired
 # , validators=[DataRequired()]
 class NewSound(FlaskForm):
@@ -7,7 +7,7 @@ class NewSound(FlaskForm):
     name = StringField("Name")
     owner_id = IntegerField("owner_id")
     is_public = BooleanField("Allow Public to View?")
-    target_volume = DecimalField("Set Volume")
-    fade_speed = DecimalField("Set the speed to fade in and out")
+    target_volume = IntegerField("Set Volume")
+    fade_speed = IntegerField("Set the speed to fade in and out")
     arrangement = IntegerField("Arrangement")
     is_looped = BooleanField("Loop Sound?")
